Update CodecademyOrganism.js
Sloppy organization. But accomplishes all tasks in the codecademy 'Mysterious Organism' exercise.

Will: 
create an array of 30 'viable' lifeforms (based on provided info)
convert DNA to RNA
extract RNA Codons (to each lifeform)
provide an array of amino acids for organism.
compare every organisms DNA with each other in an array, and return the two forms with the most in common.


DNA/RNA/amino information done to the best of my ability.